feat(keys): Explicitly track timestamp of last key rotation.
Up until now we have been using `verifierSetAt` as a default value
for `keysChangedAt`, which works but can falsely indicate that keys
have changed when in fact they have not. To support the "durable sync"
project we need to accurately track the timestamp at which the keys
last changed.

With this commit, we track a separate timestamp that is only bumped
when kB actually changes. This value will be used for OAuth scoped
key ids and reported in BrowserID assertions where it can be accessed
by the sync tokenserver.

Safe deployment of this change depends on prep work done in previous
commits. We need to be sure that sync clients all see a consistent
key-rotation timestamp regardless of whether they are using BrowserID
or OAuth, regardless of whether they hit a server running the old code
or the new code, and in the face of a rollback. This is assured as
follows:

* As of #3060, both BrowserID and OAuth clients will report the value
  from the `keysChangedAt` column if it exists, or fallback to using
  `verifierSetAt` if it is NULL.
* As of #2571, users who reset their password on a server running the
  old code will have `keysChangedAt` reset to `NULL`. In this case
  both new and old codebases will fall back to using  `verifierSetAt`.

Fixes #490.